chore: make it possible to read version info from env
