+++ ui/src/components/header/version-info/version-info.module.scss
